Add epoch suffix to Deployment object to gracefully introduce breaking changes
This change set introduces a `-v1` *epoch* suffix to the `Deployment`
object name. It allows to gracefully make breaking changes w.r.t. the
k8s API and with that changing that suffix force a re-creation of the
Deployment. An example for this is changing an immutable API field, like
what accidentially happened in
https://gitlab.com/gitlab-org/charts/gitlab-agent/-/issues/38.

This should be a "last resort" action to take and nothing to do lightly.
A proper way to introduce ANY breaking changes is still with a major
GitLab release.

Closes https://gitlab.com/gitlab-org/charts/gitlab-agent/-/issues/38