Fail publish job on release pipeline failure
The publish job silently failed in
https://gitlab.com/gitlab-org/charts/gitlab-agent/-/jobs/6486482269.

The distribution pipeline was temporarily broken which has gone
unnoticed in our publish job.

As a follow-up we should probably consider triggering a proper child
pipeline so that we can easily depend on its status.

Today, even if the pipeline fails, our publish job will still pass.

Refs https://gitlab.com/gitlab-org/charts/gitlab-agent/-/issues/55